Organized the file system
