+++ src/Components/Book/Book.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Book() {
+  return (
+    <div>Book</div>
+  )
+}
+
+export default Book
